replace Getters and Setters with lombok annotations
