+++ src/main/java/net/ictcampus/lumen_backend/controller/repositories/BlogRepository.java
@@ -1,6 +1,6 @@
-package controller.repositories;
+package net.ictcampus.lumen_backend.controller.repositories;
 
-import models.Blog;
+import net.ictcampus.lumen_backend.models.Blog;
 import org.springframework.data.repository.CrudRepository;
 
 public interface BlogRepository extends CrudRepository<Blog, String> {
