+++ src/main/java/net/ictcampus/lumen_backend/controller/services/CommentService.java
@@ -1,7 +1,7 @@
-package controller.services;
+package net.ictcampus.lumen_backend.controller.services;
 
-import controller.repositories.CommentRepository;
-import models.Comment;
+import net.ictcampus.lumen_backend.controller.repositories.CommentRepository;
+import net.ictcampus.lumen_backend.models.Comment;
 import org.springframework.stereotype.Service;
 
 @Service
